+++ test/add_test.dart
@@ -6,6 +6,7 @@
 
 import 'dart:async';
 import 'package:test/test.dart';
+import 'package:dart_config/default_server.dart' as config_file;
 
 import 'package:dartdap/dartdap.dart';
 
@@ -63,14 +64,18 @@ Future purgeEntries(LDAPConnection ldap) async {
 //----------------------------------------------------------------
 
 void doTests(String configName) {
-  var ldapConfig;
   var ldap;
 
   //----------------
 
   setUp(() async {
-    ldapConfig = new LDAPConfiguration.fromFile(testConfigFile, configName);
-    ldap = await ldapConfig.getConnection();
+    var c = (await config_file.loadConfig(testConfigFile))[configName];
+    ldap = new LDAPConnection(
+        c["host"], c["port"], c["ssl"], c["bindDN"], c["password"]);
+
+    ldap = await ldap.connect();
+    await ldap.bind();
+
     await purgeEntries(ldap);
     // Nothing to populate, since these tests exercise the "add" operation
   });
@@ -79,7 +84,7 @@ void doTests(String configName) {
 
   tearDown(() async {
     await purgeEntries(ldap);
-    await ldapConfig.close();
+    await ldap.close();
   });
 
   //----------------