+++ test/load_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:async';
 
+import 'package:dart_config/default_server.dart' as config_file;
 import 'package:test/test.dart';
 
 import 'package:dartdap/dartdap.dart';
@@ -65,14 +66,18 @@ Future purgeEntries(LDAPConnection ldap) async {
 //----------------------------------------------------------------
 
 void doTests(String configName) {
-  var ldapConfig;
   var ldap;
 
   //----------------
 
   setUp(() async {
-    ldapConfig = new LDAPConfiguration.fromFile(testConfigFile, configName);
-    ldap = await ldapConfig.getConnection();
+    var c = (await config_file.loadConfig(testConfigFile))[configName];
+    ldap = new LDAPConnection(
+        c["host"], c["port"], c["ssl"], c["bindDN"], c["password"]);
+
+    await ldap.connect();
+    await ldap.bind();
+
     await purgeEntries(ldap);
     await populateEntries(ldap);
   });
@@ -81,7 +86,7 @@ void doTests(String configName) {
 
   tearDown(() async {
     await purgeEntries(ldap);
-    await ldapConfig.close();
+    await ldap.close();
   });
 
   //----------------